requirements.txt
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cramjam==2.9.1
debugpy==1.8.11
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastparquet==2024.11.0
Flask==3.1.0
fsspec==2024.12.0
fuzzywuzzy==0.18.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
haversine==2.9.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Levenshtein==0.26.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pip==24.3.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
python-dateutil== 2.9.0.post0
pytz== 2024.2
pyzmq==26.2.0
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
rsa==4.9
scipy==1.15.0
setuptools==65.5.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3